pyproject.toml
[tool.poetry]
name = "leads-updater"
version = "0.1.0"
description = ""
authors = ["Shel Soloa <shel@shelsoloa.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
instagrapi = "^2.1.3"
pillow = "^11.1.0"
httpx = "^0.28.1"
ipdb = "^0.13.13"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
